refactor: BankAbility를 Task와 CustomerConstant로 나누고 로직에 적용
