+++ BankManagerConsoleApp/BankManagerConsoleApp/BankSystem/CustomerConstant.swift
@@ -1,0 +1,33 @@
+//
+//  CustomerConstant.swift
+//  BankManagerConsoleApp
+//
+//  Created by 송선진 on 2023/02/27.
+//
+
+import Foundation
+
+
+
+enum CustomerConstant {
+    
+    static let min: UInt = 10
+    static let max: UInt = 30
+    
+}
+
+
+
+
+extension CustomerConstant {
+    
+    //MARK: - Properties
+    
+    static var numberOfCustomer: UInt {
+        guard let customerSum = (CustomerConstant.min...CustomerConstant.max).randomElement() else { return 0 }
+        return customerSum
+    }
+    
+    
+}
+
