+++ BankManagerConsoleApp/BankManagerConsoleApp/Protocol/BankProtocol.swift
@@ -11,6 +11,6 @@ protocol BankProtocol {
     
     func open()
     func close()
-    func report(waitingNumber: UInt, taskType: BankAbility.taskType, inProgress: Bool)
+    func report(waitingNumber: UInt, taskType: Task, inProgress: Bool)
     
 }
